fix travis build error
