+++ package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node dist/server.js",
+    "test:start": "cross-env NODE_ENV=test node db.js",
     "test": "cross-env NODE_ENV=test nyc mocha server/**/*.test.js --require babel-register --exit",
     "heroku-postbuild": "babel server -d dist",
     "build": "babel server -d dist",
